changed class name with Camel  case
