+++ src/main/scala/vaquita/VecMemIO.scala
