+++ src/main/scala/vaquita/WBStage.scala
@@ -3,7 +3,7 @@ import chisel3._
 import chisel3.util._
 
 
-class wb_stage(implicit val config: Vaquita_Config,val on : Bool =1.B, val off : Bool =0.B) extends Module {
+class WBStage(implicit val config: VaquitaConfig,val on : Bool =1.B, val off : Bool =0.B) extends Module {
   val io = IO (new Bundle {
     val wb_vsd_data_in = Input(Vec(8, Vec(config.count_lanes, SInt(config.XLEN.W))))
     val wb_vs3_data_in_store = Input(Vec(8, Vec(config.count_lanes, SInt(config.XLEN.W))))
