+++ src/test/scala/vaquita/VecTopTest.scala
@@ -1,0 +1,20 @@
+package vaquita
+import chisel3._
+import chisel3.tester._
+import org.scalatest.FreeSpec
+import chisel3.experimental.BundleLiterals._
+
+class VecTopTest extends FreeSpec with ChiselScalatestTester {
+  "vec top test" in {
+    implicit val config = new VaquitaConfig {}
+    test(new VecTop) { dut =>
+    dut.io.instr.poke(0.U)
+    dut.io.rs1_data.poke(0.S)
+    dut.io.hazard_rs1_data_in.poke(0.U)
+    dut.io.vl_rs1_out.expect(0.U)    
+    dut.io.dmemReq.ready.poke(true.B)
+    dut.io.dmemRsp.valid.poke(true.B)
+    dut.clock.step(1)
+    }
+  }
+}
